cleaning up a million warnings

+++ WinDirStat/windirstat/COM_helpers.h
@@ -44,9 +44,9 @@ struct CFileDialogEventHandler final : public IFileDialogEvents, public IFileDia
 		static const QITAB qit[ ] =
 			{
 			//QITABENT( CFileDialogEventHandler, IFileDialogEvents ),
-				{ &__uuidof(IFileDialogEvents), static_cast<int>( OFFSETOFCLASS( IFileDialogEvents, CFileDialogEventHandler ) ) },
+				{ &__uuidof(IFileDialogEvents), OFFSETOFCLASS( IFileDialogEvents, CFileDialogEventHandler ) },
 			//QITABENT( CFileDialogEventHandler, IFileDialogControlEvents ),
-				{ &__uuidof(IFileDialogControlEvents), static_cast<int>( OFFSETOFCLASS( IFileDialogControlEvents, CFileDialogEventHandler ) ) },
+				{ &__uuidof(IFileDialogControlEvents), OFFSETOFCLASS( IFileDialogControlEvents, CFileDialogEventHandler ) },
 			{ 0 }
 #pragma warning( suppress: 4365)
 			};
